Remove img->origin = 1 code. The sim will properly flip the image so that the origin is 0 (top-left). I also had to change the direction of the path code, based on the angle.
Path angle returned from image: negative if /, positive if \

Therefore, the sub should turn LEFT by the angle returned. ie, if \, then sub should turn left, and if /, the sub should turn right.

This should mean that the control/vision code works the same on images given by the sim, webcam or videos. Previously, the sim images would be filtered with a vertical flip, changing the value of the angles returned of lines.